engine/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
boto3==1.39.12
botocore==1.39.12
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.4
dill==0.4.0
ecdsa==0.19.1
et_xmlfile==2.0.0
flake8==7.3.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-genai==1.25.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hvac==2.3.0
hyperframe==6.1.0
idna==3.10
inflate64==1.0.3
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
multivolumefile==0.2.3
mypy_extensions==1.1.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.11
py7zr==1.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
pylint==3.3.7
PyMuPDF==1.26.3
pypandoc==1.15
PyPDF2==3.0.1
pypdfium2==4.30.0
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-pptx==1.0.2
pytz==2025.2
pyzstd==0.17.0
RapidFuzz==3.13.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
sendgrid==6.12.4
six==1.17.0
slack_sdk==3.36.0
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
texttable==1.7.0
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xlsxwriter==3.2.9
